// src/app/Interfaces/maestro.ts
export interface Maestro {
    apellidopaterno: string;
    apellidomaterno: string;
    nombre: string;
    fechadenacimiento: string;
    edad: string;
    correo: string;
    telefono: string;
    direccion: string;
    sexo: number;
    tituloprofesional: string;
    cedulaprofesional: string;
    institucioneducativa: string; 
    tipo: number; 
}